Moving GetMatchingNodes API to search view model
